pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "environs[django]>=15.0.1",
]

[dependency-groups]
dev = [
    "django-stubs>=6.0.3",
    "mypy>=1.20.2",
    "ruff>=0.15.12",
]


[tool.mypy]
# Ignores the "Skipping analyzing" errors
ignore_missing_imports = true

# Stops the "Need type annotation" errors in models
check_untyped_defs = false
disallow_untyped_defs = false

# Tells MyPy to ignore errors from third-party libraries

plugins = [
    "mypy_django_plugin.main",
]
mypy_path = "backend"

[tool.django-stubs]
django_settings_module = "config.settings"
